Added files via upload
Started on getting set up in d3.js. Was able to get a working bar chart loading in from an external csv file. Would be nice if we could make the code a function that we can just feed the various data sets into, and implement it as an external module.